Fix a potential div by zero error in noiseFingerprint.m
If a signal was zero during a timeslot, the average value will be zero,
which can lead to a division by zero.  This is unlikely in the real
world, but happened in our evaluation, as we had to pad an interrupted
recording with zeroes.  If the average is detected as zero, it is set
to 0.00001, which should be functionally identical for most calculations,
as in this area, the relevant threshold is the absolute change, not the
relative change, either way.

Incremented version number of script to 1.1.